Remove issue tracker
After some thought, I decided that it would be better to have use a
seperate service for the issue tracker. This make it easier to add
complexity to this project and the issue tracker as each project can
specialize.

In the future, it will be possible to add hooks to
valentine, which would allow things like creating a project on the issue
tracker when a project is created on valentine.